refactor: move module old architecture model.
