+++ backend/api/build.gradle.kts
@@ -22,7 +22,7 @@ dependencies {
     implementation("org.springdoc:springdoc-openapi-starter-webmvc-api:2.4.0")
     implementation(project(":backend:applications"))
     implementation(project(":backend:core"))
-    implementation(project(":backend:domains"))
+    implementation(project(":backend:deprecateddomains"))
     implementation(project(":backend:infrastructure"))
     implementation("software.amazon.awssdk:cognitoidentity:2.25.11")
     implementation("software.amazon.awssdk:cognitoidentityprovider:2.25.11")
